02-use-cases/A2A-multi-agent-incident-response/web_search_openai_agentsclaude/scripts/utils.py
"""Import shared utilities from root scripts directory"""

import sys
from pathlib import Path

sys.path.insert(0, str(Path(__file__).parent.parent.parent / "scripts"))

from shared_utils import get_ssm_parameter, get_aws_info, invoke_endpoint, get_m2m_token_for_agent

__all__ = ["get_ssm_parameter", "get_aws_info", "invoke_endpoint", "get_m2m_token_for_agent"]
