02-use-cases/A2A-multi-agent-incident-response/web_search_openai_agentsclaude/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
a2a-sdk==0.3.25
    # via web-search-claude-agent (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
bedrock-agentcore==1.4.8
    # via web-search-claude-agent (pyproject.toml)
boto3==1.42.77
    # via
    #   bedrock-agentcore
    #   strands-agents
botocore==1.42.77
    # via
    #   bedrock-agentcore
    #   boto3
    #   s3transfer
    #   strands-agents
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via uvicorn
cryptography==46.0.6
    # via
    #   google-auth
    #   pyjwt
docstring-parser==0.17.0
    # via strands-agents
exceptiongroup==1.3.1
    # via anyio
google-api-core==2.30.0
    # via a2a-sdk
google-auth==2.49.1
    # via google-api-core
googleapis-common-protos==1.73.1
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   a2a-sdk
    #   mcp
    #   tavily-python
httpx-sse==0.4.3
    # via
    #   a2a-sdk
    #   mcp
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.1
    # via opentelemetry-api
jmespath==1.1.0
    # via
    #   boto3
    #   botocore
jsonschema==4.26.0
    # via
    #   mcp
    #   strands-agents
jsonschema-specifications==2025.9.1
    # via jsonschema
mcp==1.26.0
    # via strands-agents
opentelemetry-api==1.40.0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-threading
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   strands-agents
opentelemetry-instrumentation==0.61b0
    # via opentelemetry-instrumentation-threading
opentelemetry-instrumentation-threading==0.61b0
    # via strands-agents
opentelemetry-sdk==1.40.0
    # via strands-agents
opentelemetry-semantic-conventions==0.61b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==26.0
    # via opentelemetry-instrumentation
proto-plus==1.27.2
    # via google-api-core
protobuf==6.33.6
    # via
    #   a2a-sdk
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   a2a-sdk
    #   bedrock-agentcore
    #   mcp
    #   pydantic-settings
    #   strands-agents
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via mcp
pyjwt==2.12.1
    # via mcp
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.2.2
    # via pydantic-settings
python-multipart==0.0.22
    # via mcp
pyyaml==6.0.3
    # via strands-agents
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2026.2.28
    # via tiktoken
requests==2.33.0
    # via
    #   google-api-core
    #   tavily-python
    #   tiktoken
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sse-starlette==3.3.3
    # via mcp
starlette==1.0.0
    # via
    #   bedrock-agentcore
    #   mcp
    #   sse-starlette
strands-agents==1.33.0
    # via web-search-claude-agent (pyproject.toml)
tavily-python==0.7.23
    # via web-search-claude-agent (pyproject.toml)
tiktoken==0.12.0
    # via tavily-python
typing-extensions==4.15.0
    # via
    #   anyio
    #   bedrock-agentcore
    #   cryptography
    #   exceptiongroup
    #   mcp
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   pyjwt
    #   referencing
    #   starlette
    #   strands-agents
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via
    #   bedrock-agentcore
    #   botocore
    #   requests
uvicorn==0.42.0
    # via
    #   web-search-claude-agent (pyproject.toml)
    #   bedrock-agentcore
    #   mcp
watchdog==6.0.0
    # via strands-agents
websockets==16.0
    # via bedrock-agentcore
wrapt==1.17.3
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-threading
zipp==3.23.0
    # via importlib-metadata
